refactor: change test file names and remove unneccesary e2e tests
