+++ e2e-tests/addperson.spec.ts
@@ -19,32 +19,11 @@ async function fillRequiredFields(page: Page) {
   await page.getByLabel("Sopimuksen loppu:").fill("2026-01-01");
 }
 
-test("person modal can be opened", async ({ page }) => {
+test("add person modal can be opened", async ({ page }) => {
   await openAddPersonModal(page);
   await expect(
     page.getByRole("heading", { name: "Lisää henkilö" }),
   ).toBeVisible();
-});
-
-test("close button opens confirmation dialog", async ({ page }) => {
-  await openAddPersonModal(page);
-  await page.locator(".personmodal-close-button").click();
-  await expect(page.getByText("Sulje ilman tallennusta?")).toBeVisible();
-});
-
-test("clicking outside modal opens confirmation dialog", async ({ page }) => {
-  await openAddPersonModal(page);
-  await page.locator(".personmodal-overlay").dispatchEvent("click");
-  await expect(page.getByText("Sulje ilman tallennusta?")).toBeVisible();
-});
-
-test("confirming close via overlay dismisses modal", async ({ page }) => {
-  await openAddPersonModal(page);
-  await page.locator(".personmodal-overlay").dispatchEvent("click");
-  await page.getByText("Kyllä").click();
-  await expect(
-    page.getByRole("heading", { name: "Lisää henkilö" }),
-  ).not.toBeVisible();
 });
 
 test("confirming close dismisses modal", async ({ page }) => {
@@ -82,16 +61,6 @@ test("save button is enabled when required fields are filled", async ({
   ).toBeEnabled();
 });
 
-test("dropdowns can be selected", async ({ page }) => {
-  await openAddPersonModal(page);
-  await page.getByLabel("Osasto:").selectOption({ index: 1 });
-  await page.getByLabel("Työnimike:").selectOption({ index: 1 });
-  await page.getByLabel("Tutkimusryhmä:").selectOption({ index: 1 });
-  await expect(page.getByLabel("Osasto:")).not.toHaveValue("");
-  await expect(page.getByLabel("Työnimike:")).not.toHaveValue("");
-  await expect(page.getByLabel("Tutkimusryhmä:")).not.toHaveValue("");
-});
-
 test("saving opens confirmation dialog", async ({ page }) => {
   await openAddPersonModal(page);
   await fillRequiredFields(page);
@@ -100,6 +69,12 @@ test("saving opens confirmation dialog", async ({ page }) => {
     .getByRole("button", { name: "Lisää", exact: true })
     .click();
   await expect(page.getByText("Tallenna muutokset?")).toBeVisible();
+});
+
+test("clicking outside modal opens confirmation dialog", async ({ page }) => {
+  await openAddPersonModal(page);
+  await page.locator(".personmodal-overlay").dispatchEvent("click");
+  await expect(page.getByText("Sulje ilman tallennusta?")).toBeVisible();
 });
 
 test("cancelling save confirmation keeps modal open", async ({ page }) => {
@@ -115,7 +90,7 @@ test("cancelling save confirmation keeps modal open", async ({ page }) => {
   ).toBeVisible();
 });
 
-test("confirming save closes modal and persists person", async ({ page }) => {
+test("confirming save closes modal", async ({ page }) => {
   await openAddPersonModal(page);
   await fillRequiredFields(page);
   await page
